% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{coordToTanV}
\alias{coordToTanV}
\alias{coordToTanV.default}
\title{Transform the coordinates for the tangent space into a tangent vector}
\usage{
coordToTanV(mfd, p, V0, ...)

\method{coordToTanV}{default}(mfd, p, V0, ...)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{p}{A vector for the base point}

\item{V0}{A matrix or a vector containing the coordinates}

\item{...}{Passed into specific methods}
}
\value{
A matrix in which each column correspond to a different tangent vector
}
\description{
Transform the coordinates w.r.t. a basis returned by [basisTan()] to tangent vectors
}
\section{Methods (by class)}{
\itemize{
\item \code{coordToTanV(default)}: Method

}}
