% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExpM}
\alias{ExpM}
\title{Matrix exponential}
\usage{
ExpM(X)
}
\arguments{
\item{X}{A square matrix}
}
\value{
A matrix having the same size as the input
}
\description{
\code{ExpM(X)} computes the matrix exponential using Armardillo.
}
