% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{is.finiteDim}
\alias{is.finiteDim}
\title{Tell whether a manifold is a finite-dimensional one}
\usage{
is.finiteDim(mfd)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}
}
\value{
A logical scalar indicating whether the manifold is finite-dimensional
}
\description{
Tell whether a manifold is a finite-dimensional one
}
