% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{listAvailMfd}
\alias{listAvailMfd}
\title{List all available manifold names}
\usage{
listAvailMfd()
}
\value{
A character vector of the names of available manifolds
}
\description{
List all available manifold names
}
