% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDManifolds.R
\name{MakeSym}
\alias{MakeSym}
\title{Make a symmetric matrix by specifying a near-symmetric matrix M, or the lower triangular elements lowerTri with diagonal.}
\usage{
MakeSym(M, lowerTri, doubleDiag = FALSE)
}
\arguments{
\item{M}{A near-symmetric matrix}

\item{lowerTri}{A vector containing the lower triangular elements of the matrix. This is an alternative way to specify the matrix.}

\item{doubleDiag}{Only meaningful for lowerTri is not missing. Whether the diagonal elements should be multiplied by sqrt(2) for doubling the squared norm of the lower triangle}
}
\value{
A symmetric matrix
}
\description{
Make a symmetric matrix by specifying a near-symmetric matrix M, or the lower triangular elements lowerTri with diagonal.
}
