% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{manifold}
\alias{manifold}
\title{manifold: Operations for Riemannian manifolds}
\description{
Implements operations for Riemannian manifolds, e.g., geodesic distance, Riemannian metric, exponential and logarithm maps, etc. Also incorporates random object generator on the manifolds
}
\details{
References: 
Dai X, Lin Z, Müller HG. Modeling sparse longitudinal data on Riemannian manifolds. Biometrics. 2021;77(4):1328–41. 
Dai X, Lopez-Pintado S. Tukey’s depth for object data. Journal of the American Statistical Association. 2021;In press. 


Maintainer:  Xiongtao Dai \email{xdai@berkeley.edu}
}
\author{
Xiongtao Dai \email{xdai@berkeley.edu}
Zhenhua Lin
}
