% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Normalize}
\alias{Normalize}
\title{Normalize a vector}
\usage{
Normalize(v, tol = 1e-10)
}
\arguments{
\item{v}{A vector to be normalized to have a unit norm}

\item{tol}{The tolerance value for deciding the v is actually 0, cannot be normalized, and thus return just the 0 vector}
}
\value{
A vector with the same length as `v` but having unit Euclidean norm
}
\description{
Normalize a vector
}
