% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuManifold.R, R/General.R, R/HSManifold.R,
%   R/SOManifold.R, R/SPDManifolds.R, R/SphereManifold.R
\name{project.Euclidean}
\alias{project.Euclidean}
\alias{project}
\alias{project.L2}
\alias{project.HS}
\alias{project.Dens}
\alias{project.SO}
\alias{project.AffInv}
\alias{project.LogEu}
\alias{project.SPD}
\alias{project.Sphere}
\title{Project data points in the ambient space onto the manifold}
\usage{
\method{project}{Euclidean}(mfd, p)

project(mfd, p)

\method{project}{L2}(mfd, p)

\method{project}{HS}(mfd, p)

\method{project}{Dens}(mfd, p)

\method{project}{SO}(mfd, p)

\method{project}{AffInv}(mfd, p)

\method{project}{LogEu}(mfd, p)

\method{project}{SPD}(mfd, p)

\method{project}{Sphere}(mfd, p)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{p}{A matrix of points to be projected to the manifold.}
}
\value{
A matrix in which columns correspond to points on the manifold
}
\description{
Project data points in the ambient space onto the manifold
}
\section{Methods (by class)}{
\itemize{
\item \code{project(Euclidean)}: Method

\item \code{project(L2)}: Method

\item \code{project(HS)}: Method

\item \code{project(Dens)}: Method

\item \code{project(SO)}: Method

\item \code{project(AffInv)}: Method

\item \code{project(LogEu)}: Method

\item \code{project(SPD)}: Method

\item \code{project(Sphere)}: Method

}}
