% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{tanVToCoord}
\alias{tanVToCoord}
\alias{tanVToCoord.default}
\title{Transform tangent vectors to their coordinates on the tangent space}
\usage{
tanVToCoord(mfd, p, V, ...)

\method{tanVToCoord}{default}(mfd, p, V, ...)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{p}{A vector for the base point}

\item{V}{A matrix or a vector containing the tangent vectors}

\item{...}{Passed into specific methods}
}
\value{
A matrix in which each column correspond to a coordinate
}
\description{
Transform the tangent vectors to coordinates w.r.t. the basis returned by [basisTan()]
}
\section{Methods (by class)}{
\itemize{
\item \code{tanVToCoord(default)}: Method

}}
